chore: Rename to SMS campaign
